+++ Sudoku/src/com/Dave/Sudoku/ScoringVanilla.java
@@ -4,12 +4,12 @@ import android.graphics.Point;
 
 public class ScoringVanilla implements IScoring
 {
-	public int ScoreMove(int[][] fullBoard, Point point, int number, int multiplier)
+	public int ScoreMove(byte[][] fullBoard, Point point, byte number, int multiplier)
 	{
 		return number;
 	}
 	
-	public int GetNextMultiplier(int[][] fullBoard, Point point, int number)
+	public int GetNextMultiplier(byte[][] fullBoard, Point point, byte number)
 	{
 		return 0;
 	}
